refactor: separate country list
