+++ performance/src/components/country-list.tsx
@@ -1,12 +1,12 @@
-import { ArrowDownUp, ArrowDownWideNarrow, ArrowUpWideNarrow } from 'lucide-react';
 import { useState } from 'react';
-import { List } from 'react-virtualized';
 
 import { useCountries } from '@/hooks/use-countries';
 import type { SortConfig } from '@/types/co2-data';
 import { cn } from '@/utils/cn';
 
-import SearchInput from './search-input';
+import { CountryTable } from './country-table';
+import { TableHeader } from './table-header';
+import { TableToolbar } from './table-toolbar';
 
 function CountryList() {
   const countriesData = useCountries();
@@ -21,15 +21,6 @@ function CountryList() {
       key,
       direction: prev.key === key && prev.direction === 'asc' ? 'desc' : 'asc',
     }));
-  };
-
-  const getSortIcon = (key: string) => {
-    if (sortConfig.key !== key) return <ArrowDownUp className='mx-2 w-5 h-5'/>;
-    return sortConfig.direction === 'asc' ? (
-      <ArrowUpWideNarrow className="mx-2 h-5 w-5" />
-    ) : (
-      <ArrowDownWideNarrow className="mx-2 h-5 w-5" />
-    );
   };
 
   const filteredCountries = Object.entries(countriesData)
@@ -76,34 +67,6 @@ function CountryList() {
     return 0;
   });
 
-   const rowRenderer = ({
-     index,
-     key,
-     style,
-   }: {
-     index: number;
-     key: string;
-     style: React.CSSProperties;
-   }) => {
-     const country = sortedCountries[index];
-     return (
-       <div
-         key={key}
-         style={style}
-         className={cn(
-           'even:bg-input/20 w-full',
-           'grid grid-cols-[2fr_1fr_1fr_1fr]',
-           'gap-4 px-3 py-2 odd:bg-transparent',
-         )}
-       >
-         <div>{country.name}</div>
-         <div>{country.isoCode || 'N/A'}</div>
-         <div>{country.year || 'N/A'}</div>
-         <div>{country.population ? country.population.toLocaleString() : 'N/A'}</div>
-       </div>
-     );
-   };
-
   return (
     <div
       className={cn(
@@ -114,53 +77,15 @@ function CountryList() {
         'custom-scrollbar overflow-y-auto',
       )}
     >
-      <div className="bg-input flex w-full items-center justify-between px-2 py-2">
-        <h2 className="justify-center px-3 pt-2 text-center text-lg font-semibold">
-          Countries ({sortedCountries.length})
-        </h2>
-        <SearchInput value={searchTerm} onChange={setSearchTerm} placeholder="Search country..." />
-      </div>
-      <div
-        className={cn(
-          'bg-input border-shadow/20 w-full',
-          'grid grid-cols-[2fr_1fr_1fr_1fr]',
-          'border-b font-semibold sm:gap-4',
-        )}
-      >
-        <button
-          onClick={() => handleSort('name')}
-          className="hover:bg-input/30 px-5 py-1 text-left font-bold duration-300 hover:scale-103 sm:flex"
-        >
-          Country {getSortIcon('name')}
-        </button>
-        <button
-          onClick={() => handleSort('isoCode')}
-          className="hover:bg-input/30 px-5 py-1 text-left font-bold whitespace-nowrap duration-300 hover:scale-103 sm:flex"
-        >
-          ISO Code {getSortIcon('isoCode')}
-        </button>
-        <button
-          onClick={() => handleSort('year')}
-          className="hover:bg-input/30 px-5 py-1 text-left font-bold duration-300 hover:scale-103 sm:flex"
-        >
-          Year {getSortIcon('year')}
-        </button>
-        <button
-          onClick={() => handleSort('population')}
-          className="hover:bg-input/30 px-5 py-1 text-left font-bold duration-300 hover:scale-103 sm:flex"
-        >
-          Population {getSortIcon('population')}
-        </button>
-      </div>
-      <div className="custom-scrollbar flex w-full flex-col overflow-y-auto">
-        <List
-          width={800} 
-          height={400} 
-          rowCount={sortedCountries.length}
-          rowHeight={40} 
-          rowRenderer={rowRenderer}
-        />
-      </div>
+      <TableToolbar
+        searchTerm={searchTerm}
+        onSearchChange={setSearchTerm}
+        countriesCount={sortedCountries.length}
+      />
+
+      <TableHeader sortConfig={sortConfig} onSort={handleSort} />
+
+      <CountryTable countries={sortedCountries} />
     </div>
   );
 }
